Refactored load methods from SceneManager into AssetManager
