+++ src/main/java/ui/AssetManager.java
@@ -1,7 +1,10 @@
 package ui;
 
+import datasource.*;
 import javafx.scene.image.Image;
+import javafx.scene.text.Font;
 
+import java.io.InputStream;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -11,6 +14,44 @@ public class AssetManager {
     private final Map<String, String> svgPaths = new HashMap<>();
 
     private String cssUrl;
+
+    public void loadGlobalFiles() {
+        loadCSS();
+        loadImages();
+        loadIcon("restart", "/icons/restart.txt");
+        loadIcon("left-bracket", "/icons/left-bracket.txt");
+
+        loadFont("/fonts/Koulen-Regular.ttf");
+        loadFont("/fonts/NationalPark-VariableFont_wght.ttf");
+    }
+
+    private void loadCSS() {
+        FileLoader loader = new StyleSheetLoader();
+        loader.open("/styles.css");
+        String cssUrl = loader.getFileUrl().toExternalForm();
+        setStylesheet(cssUrl);
+    }
+
+    private void loadImages() {
+        FileLoader loader = new ImageLoader();
+        loader.open("/images/placeholder.png");
+        String imageUrl = loader.getFileUrl().toExternalForm();
+        addImage("placeholder", imageUrl);
+    }
+
+    private void loadIcon(String key, String fileName) {
+        IconLoader loader = new IconLoader();
+        loader.open(fileName);
+        String pathData = loader.getPathData();
+        addSvg(key, pathData);
+    }
+
+    private void loadFont(String fileName) {
+        FileLoader loader = new FontLoader();
+        loader.open(fileName);
+        InputStream fontStream = getClass().getResourceAsStream(fileName);
+        Font.loadFont(fontStream, 12);
+    }
 
     public void addImage(String key, String imageUrl) {
         Image image = new Image(imageUrl);
