+++ src/main/java/ui/SceneManager.java
@@ -12,15 +12,17 @@ import java.io.InputStream;
 public class SceneManager {
 
     private final Stage stage;
-    private final GameState model = new GameState();
-    private final AssetManager assets = new AssetManager();
+    private final GameState model;
+    private final AssetManager assets;
 
     public SceneManager(Stage stage) {
         this.stage = stage;
+        this.model = new GameState();
+        this.assets = new AssetManager();
     }
 
     public void showPlayerDeckView() {
-        loadGlobalFiles();
+        assets.loadGlobalFiles();
         PlayerDeckController controller = new PlayerDeckController(model, assets);
         setScene(controller.getView().getRoot());
     }
@@ -36,44 +38,6 @@ public class SceneManager {
         stage.setResizable(false);
 
         stage.show();
-    }
-
-    private void loadGlobalFiles() {
-        loadCSS();
-        loadImages();
-        loadIcon("restart", "/icons/restart.txt");
-        loadIcon("left-bracket", "/icons/left-bracket.txt");
-
-        loadFont("/fonts/Koulen-Regular.ttf");
-        loadFont("/fonts/NationalPark-VariableFont_wght.ttf");
-    }
-
-    private void loadCSS() {
-        FileLoader loader = new StyleSheetLoader();
-        loader.open("/styles.css");
-        String cssUrl = loader.getFileUrl().toExternalForm();
-        assets.setStylesheet(cssUrl);
-    }
-
-    private void loadImages() {
-        FileLoader loader = new ImageLoader();
-        loader.open("/images/placeholder.png");
-        String imageUrl = loader.getFileUrl().toExternalForm();
-        assets.addImage("placeholder", imageUrl);
-    }
-
-    private void loadIcon(String key, String fileName) {
-        IconLoader loader = new IconLoader();
-        loader.open(fileName);
-        String pathData = loader.getPathData();
-        assets.addSvg(key, pathData);
-    }
-
-    private void loadFont(String fileName) {
-        FileLoader loader = new FontLoader();
-        loader.open(fileName);
-        InputStream fontStream = getClass().getResourceAsStream(fileName);
-        Font.loadFont(fontStream, 12);
     }
 
 }
